Move useSearchAccounts into /hooks as well
