+++ app/javascript/mastodon/hooks/useSearchAccounts.ts
